backend/.requirements.txt
wheel
psycopg2
python-dotenv
django
djangorestframework
django-cors-headers
django_ratelimit
gunicorn
